+++ moabb/evaluations/__init__.py
@@ -9,4 +9,5 @@ from .evaluations import (
     CrossSubjectEvaluation,
     WithinSessionEvaluation,
 )
+from .splitters import WithinSessionSplitter
 from .utils import create_save_path, save_model_cv, save_model_list